+++ src/Facades/Socialite.php
@@ -7,6 +7,7 @@ use Laravel\Socialite\Contracts\Factory;
 
 /**
  * @method static \Laravel\Socialite\Contracts\Provider driver(string $driver = null)
+ * @method static \Laravel\Socialite\Two\AbstractProvider buildProvider($provider, $config)
  *
  * @see \Laravel\Socialite\SocialiteManager
  */