+++ src/Two/GithubProvider.php
@@ -90,6 +90,7 @@ class GithubProvider extends AbstractProvider implements ProviderInterface
     {
         return (new User)->setRaw($user)->map([
             'id' => $user['id'],
+            'nodeId' => $user['node_id'],
             'nickname' => $user['login'],
             'name' => Arr::get($user, 'name'),
             'email' => Arr::get($user, 'email'),